requirement.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0

phonenumbers==8.13.2
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3